change farm classes for another protocols
