+++ flag_senders/errors.py
@@ -1,0 +1,8 @@
+class FarmError(Exception):
+	pass
+
+class CliError(Exception):
+	pass
+
+class ProtocolError(Exception):
+	pass
